% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaseControl.R
\name{createCaseControlAnalysesDetails}
\alias{createCaseControlAnalysesDetails}
\title{Create the analyses details}
\usage{
createCaseControlAnalysesDetails(outputFolder)
}
\arguments{
\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Create the analyses details
}
\details{
This function creates files specifying the analyses that will be performed.
}
