% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFiguresAndTables.R
\name{createFiguresAndTables}
\alias{createFiguresAndTables}
\title{Create figures and tables}
\usage{
createFiguresAndTables(connectionDetails, cdmDatabaseSchema, oracleTempSchema,
  outputFolder)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Create figures and tables for the paper.
}
