% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Icd9CovariateBuilder.R
\name{createIcd9CovariateSettings}
\alias{createIcd9CovariateSettings}
\title{Create settings for covariates representing one or more ICD-9 codes}
\usage{
createIcd9CovariateSettings(covariateDefs, windowStart = -365,
  windowEnd = -1)
}
\arguments{
\item{covariateDefs}{A data frame with three columns: covariateId, covariateName, icd9}

\item{windowStart}{Days relative to the index date to start capturing the covariates. Negative numbers indicates
days prior to index.}

\item{windowEnd}{Days relative to the index date to end capturing the covariates. Negative numbers indicates
days prior to index.}
}
\description{
Create settings for covariates representing one or more ICD-9 codes
}
