% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute OHDSI Evaluating the Case-Control Design Study}
\usage{
execute(connectionDetails, cdmDatabaseSchema, cohortDatabaseSchema, cohortTable,
  oracleTempSchema, outputFolder, createCohorts = TRUE,
  synthesizePositiveControls = TRUE, runAnalyses = TRUE,
  createFiguresAndTables = TRUE, maxCores = 4)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{createCohorts}{Instantiate the necessary cohorts in the cohortTable?}

\item{synthesizePositiveControls}{Synthesize positive controls based on the negative controls?}

\item{runAnalyses}{Execute the case-control analaysis?}

\item{createFiguresAndTables}{Create the figures and tables for the paper?}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}
}
\description{
This function executes the OHDSI Evaluating the Case-Control Design Study.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(connectionDetails,
        cdmDatabaseSchema = "cdm_data",
        cohortDatabaseSchema = "results",
        oracleTempSchema = NULL,
        outputFolder = "c:/temp/study_results")

}

}
